cleaned a little all the unused code
