+++ notan/crates/notan_draw/src/lib.rs
@@ -2,6 +2,7 @@ mod batch;
 mod builder;
 mod custom_pipeline;
 mod draw;
+mod extension;
 mod images;
 mod manager;
 mod patterns;
@@ -13,6 +14,7 @@ pub mod prelude;
 
 pub use custom_pipeline::*;
 pub use draw::*;
+pub use extension::*;
 pub use images::*;
 pub use manager::*;
 pub use patterns::*;
