+++ notan/crates/notan_glyph/src/lib.rs
@@ -2,16 +2,14 @@ pub mod prelude;
 
 mod font;
 mod font_vertex;
-mod manager;
 mod owned_text;
+mod pipeline;
 mod plugin;
-mod renderer;
 mod text;
 
 pub use font::Font;
 pub use font_vertex::*;
-pub use manager::GlyphManager;
 pub use owned_text::OwnedText;
+pub use pipeline::*;
 pub use plugin::GlyphPlugin;
-pub use renderer::*;
 pub use text::Text;
