+++ notan/crates/notan_glyph/src/prelude.rs
@@ -1,6 +1,6 @@
 pub use crate::font::Font;
 pub use crate::font_vertex::*;
-pub use crate::manager::GlyphManager;
 pub use crate::owned_text::OwnedText;
-pub use crate::renderer::*;
+pub use crate::pipeline::*;
+pub use crate::plugin::GlyphPlugin;
 pub use crate::text::Text;
