+++ notan/examples/draw_triangle.rs
@@ -4,8 +4,8 @@ use notan::prelude::*;
 #[notan::main]
 fn main() -> Result<(), String> {
     notan::init_with(|gfx: &mut Graphics| {
-        let ext = DrawPlugin::new(gfx).unwrap();
-        gfx.plugins.set(ext);
+        let ext = DrawExtension::new(gfx).unwrap();
+        gfx.add_ext(ext);
     })
     .draw(draw)
     .build()
@@ -15,5 +15,5 @@ fn draw(gfx: &mut Graphics) {
     let mut draw = gfx.create_draw();
     draw.clear(Color::BLACK);
     draw.triangle((400.0, 100.0), (100.0, 500.0), (700.0, 500.0));
-    gfx.r(&draw);
+    gfx.render(&draw);
 }
